Solution to Check Profit or Loss

+++ Check Profit or Loss@@6788dd63490637cfc44e1b98/code.c
@@ -1,8 +1,8 @@
 // Your code here..
 #include<stdio.h>
 int main(){
-    int p,l;
-    scanf("%d %d",&p,&l);
+    float p,l;
+    scanf("%f %f",&p,&l);
     if(p>l){
         printf("Profit");
     }
